Add feautre of the day component
