--- routes/sightings.js
@@ -1,69 +1,0 @@
-const express = require("express");
-const router = express.Router();
-const Sighting = require("./models/sighting");
-
-// Build filter helper function
-function sightingsFilters(query, userIdString) {
-  const filters = {};
-  if (query.onlyYours == "true") {
-    filters.userId = userIdString;
-  }
-  return filters;
-}
-
-// Get sightings
-router.get("/sightings", async (req, res) => {
-  if (req.session.user) {
-    // apply filter before getting sightings
-    const filter = sightingsFilters(req.query, req.session.user._id);
-    const sightings = await Sighting.find(filter);
-    res.json(sightings);
-  } else {
-    const sightings = await Sighting.find({});
-    res.json(sightings);
-  }
-});
-
-// Get your sightings
-router.get("/yourSightings", async (req, res) => {
-  if (req.session.user) {
-    const sightings = await Sighting.find({ userId: req.session.user._id });
-    res.json(sightings);
-  } else {
-    return res.status(404).json({ error: "User not found" });
-  }
-});
-
-// Submit a sighting
-router.post("/submitSighting", async (req, res) => {
-  if (req.session.user) {
-    try {
-      const newSighting = new Sighting({
-        userId: req.session.user._id,
-        username: req.session.user.username,
-        species: req.body.species,
-        description: req.body.description || "",
-        location: {
-          type: "Point",
-          coordinates: req.body.coordinates, // [lng, lat]
-        },
-        photoUrl: req.body.photoUrl || "",
-        timestamp: new Date(req.body.timestamp),
-        taxonomicGroup: req.body.taxonomicGroup,
-        userDescription: req.body.userDescription,
-      });
-      const newSightingSaved = await newSighting.save();
-      res
-        .status(201)
-        .json({ message: "Sighting saved", data: newSightingSaved });
-    } catch (err) {
-      res.status(500).json({ error: "Failed to save sighting" });
-    }
-  } else {
-    return res.status(404).json({
-      error: "User not found. You must be logged in to submit a sighting.",
-    });
-  }
-});
-
-module.exports = router;
